Core/Auras: Initialize aura spell visual from spell that created it instead of recalculating it
